move helix_client to rocksdb_admin
